refactor(#417 W-B5, Rust half): five logging forwarders retired, and nros_trace! keeps its prefix
Deleted `nros_debug!`, `nros_info!`, `nros_warn!`, `nros_error!` and
`nros_fatal!` -- five, matching the phase doc. My brief to the sweep named only
four; `nros_fatal!` -> `log_fatal!` is the fifth.

`nros_trace!` SURVIVES, and that is the point of the change rather than an
exception to it: rclrs stops at `debug`, so TRACE has no upstream twin and there
is no `log_trace!` anywhere in the tree to rename to. The
`nros_*_throttle!` family keeps its prefix for the same reason (separate
ledgered divergence).

`tests/deprecated_forwarders.rs` went with the forwarders, but it contained
`trace_keeps_our_prefix_because_rclrs_has_no_trace` -- an assertion about the
macro that SURVIVES, sitting in a file named after the ones that do not. Moved
to `tests/trace_is_not_renamed.rs`, whose module doc now says why it outlived
its neighbours. It passes in both branches: delivering under `max-level-trace`,
compiled out under `max-level-warn`. That is exactly the coverage a
delete-the-file sweep loses without noticing.

`deprecate-legacy-names` was in `default` (W-B1 armed it) and enabled by nothing
else. Removed from `default`, definition and rationale deleted.

Three classes of leftover the armed `#[deprecated]` could not reach, all found
by building rather than grepping:

* **3 call sites in `examples/qemu-esp32-baremetal/rust/{listener,talker}`.**
  Those leaves are in Cargo.toml's `exclude` block, not `members`, so a
  workspace-wide deprecation warning never enumerates them. Same shape as
  nros-bridge two commits ago, which only the workspace lane compiles.
* **2 re-exports in the `nros` facade** (`lib.rs:290` root, `:1513` prelude).
  A `!`-anchored grep cannot see `pub use nros_log::{nros_debug, ...}`; the
  first sweep was clean and wrong, and `just check workspace-all` found them.
* **2 gate scripts naming deleted macros.** `check-no-std-stdio.py` told users
  at failure time to write `nros_log::nros_error!`, and
  `check-no-silent-sample-drop.py`'s `SAYS_SOMETHING` regex no longer matched
  `log_error`/`log_warn` -- so a migrated Rust error path would have read as a
  SILENT DROP. A gate whose remedy text names a deleted symbol is worse than
  stale: it is a wrong instruction issued at the moment someone is stuck.

Verified: `cargo test -p nros-log` green on default and on
`--no-default-features --features max-level-warn,buffer-size-256`, `log-compat`,
`std`, `alloc`; `just check fast` (197 gates) and `just check workspace-all`
both exit 0. `--features platform-clock` fails on `undefined reference to
nros_platform_clock_ns` -- pre-existing, breaks untouched targets, and the
manifest documents that extern as a link-time requirement no host build
satisfies.
