+++ packages/core/nros-log/Cargo.toml
@@ -31,7 +31,7 @@ categories = ["embedded", "no-std", "development-tools::debugging"]
 # smaller N = more truncation. Pick at most one.
 
 [features]
-default = ["max-level-trace", "buffer-size-256", "deprecate-legacy-names"]
+default = ["max-level-trace", "buffer-size-256"]
 std = ["alloc"]
 alloc = []
 
@@ -79,27 +79,6 @@ buffer-size-128  = []
 buffer-size-256  = []
 buffer-size-512  = []
 buffer-size-1024 = []
-
-# phase-417 — ARMS the `#[deprecated]` on the pre-rename `nros_debug!` ..
-# `nros_fatal!` forwarders. It WAS off by default, and the reason was structural
-# rather than timid: this workspace sets `warnings = "deny"` (root `Cargo.toml`
-# `[workspace.lints.rust]`), so an always-armed deprecation is not a migration
-# STEP -- it is a hard compile error in every crate that has not moved yet, i.e.
-# the flag day RFC-0089's two-step alias exists to avoid. Measured at the rename:
-# 9 errors in `nros-rmw-cffi` alone stopped the workspace build.
-#
-# So the attribute was authored, reviewed and inert, and the migration batch is
-# `--features deprecate-legacy-names` -> fix what it names -> delete the
-# forwarders.
-#
-# phase-417 stage 6 W-B1, 2026-09-04: THE SWITCH IS NOW ON -- it moved into
-# `default` above, which is what made the compiler enumerate the batch instead
-# of a grep guessing at it (208 macro call sites across 50 files, plus the
-# `use` lists that import them). It stays a named feature rather than being
-# deleted outright so W-B5 has exactly one thing to remove alongside the five
-# forwarders, and so a consumer mid-migration can still buy time with
-# `default-features = false` plus the level/buffer features it wants.
-deprecate-legacy-names = []
 
 # Optional `log` crate bridge (Phase 88.10). Lets ecosystem crates
 # that target `log::info!` etc. feed through the same dispatcher.
